pyproject.toml
[tool.poetry]
name = "govbrnews-mcp"
version = "0.1.0"
description = "MCP Server para notícias governamentais brasileiras via Typesense"
authors = ["Nitai Bezerra da Silva <nitaibezerra@protonmail.com>"]
readme = "README.md"
license = "MIT"
keywords = ["mcp", "typesense", "govbr", "news", "brazilian-government"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = "^3.10"
fastmcp = "^2.0"
typesense = "^0.21.0"
pydantic = "^2.0"
pydantic-settings = "^2.0"
python-dotenv = "^1.0"
cachetools = "^5.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
black = "^24.0"
ruff = "^0.3"
ipdb = "^0.13"

[tool.poetry.scripts]
govbrnews-mcp = "govbrnews_mcp.server:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"
